requirements.txt
bip-utils==1.6.0
colorama==0.4.4
ecdsa==0.16.1
Pillow==8.0.1
pypng==0.0.20
pysha3==1.0.2
python-dotenv==0.15.0
qrcode==6.1
six==1.15.0
